finish implementation of afi33
