+++ src/test/java/org/javamodularity/moduleplugin/tasks/ClasspathFileTest.java
@@ -1,11 +1,6 @@
 package org.javamodularity.moduleplugin.tasks;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertSame;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-import groovy.util.Node;
+import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
@@ -17,6 +12,8 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+
+import groovy.util.Node;
 
 /**
  * Class testing {@link ClasspathFile}.
@@ -61,179 +58,161 @@ final class ClasspathFileTest {
   } // end method */
   
   /**
-   * Test method for {@link ClasspathFile#checkGradleScopeTest(Node)}.
+   * Test method for {@link ClasspathFile#ClasspathFile(org.gradle.api.Project)}.
    */
   @Test
-  void test_checkGradleScopeTest__Node() {
+  void test_ClasspathFile__Project() {
+    fail("Not yet implemented"); // TODO
+  } // end method */
+  
+  /**
+   * Test method for {@link ClasspathFile#configure()}.
+   */
+  @Test
+  void test_configure() {
+    fail("Not yet implemented"); // TODO
+  } // end method */
+  
+  /**
+   * Test method for {@link ClasspathFile#improveEclipseClasspathFile(Node)}.
+   */
+  @Test
+  void test_improveEclipseClasspathFile__Node() {
     // Test strategy:
-    // --- a. item without children
-    // --- b. item without proper children
-    // --- c. item with "attributes" but no grand-children
-    // --- d. item -> "attributes" -> "attribute" but no attributes
-    // --- e. item -> "attributes" -> "attribute" but non gradle scope attributes
-    // --- f. item -> "attributes" -> "attribute" but gradle scope != test
-    // --- g. item -> "attributes" -> "attribute" and gradle scope == test, no  test-attribute
-    // --- h. item -> "attributes" -> "attribute" and gradle scope == test plus test-attribute
-    // --- i. only first child named "attributes" count
+    // ... assertion 1: method putJreOnModulePath(Node) works as intended
+    // --- a. JRE, rootNode with a bunch of entries and entries differing slightly
     
-    Node dut;
+    // --- a. JRE, rootNode with a bunch of entries and entries differing slightly
+    final Node rootNode = new Node(null, "root");
     
-    // --- a. item without children
-    dut = new Node(null, "item");
-    insDut.checkGradleScopeTest(dut);
+    // a.1: difference in name of item
+    final Map<String, String> mapA1 = new LinkedHashMap<>();
+    mapA1.put("path", "prefix" + "JRE_CONTAINER" + "suffix");
+    mapA1.put("kind", "con");
+    rootNode.appendNode("classpathentry", mapA1); // ok
+    rootNode.appendNode("classpathEntry", mapA1); // not classpathentry
+    
+    // --- improve
+    insDut.improveEclipseClasspathFile(rootNode);
+    
+    // --- check
     assertEquals(
-        "item[attributes={}; value=[]]",
-        dut.toString()
+        "root[attributes={}; value=["
+        // a.1, begin
+        +   "classpathentry[attributes={path=prefixJRE_CONTAINERsuffix, kind=con}; value=["
+        +     "attributes[attributes={}; value=["
+        +       "attribute[attributes={name=module, value=true}; value=[]]"
+        +     "]]]"
+        +   "], "
+        +   "classpathEntry[attributes={path=prefixJRE_CONTAINERsuffix, kind=con}; value=[]]"
+        + "]]",
+        rootNode.toString()
     );
     
-    // --- b. item without proper children
-    dut.appendNode("foo.bar");
-    insDut.checkGradleScopeTest(dut);
-    assertEquals(
-        "item[attributes={}; value=["
-        + "foo.bar[attributes={}; value=[]]"
-        + "]]",
-        dut.toString()
-    );
+    fail("Not yet implemented"); // TODO
+  } // end method */
+
+  /**
+   * Test method for {@link ClasspathFile#markMain(groovy.util.Node)}.
+   */
+  @Test
+  void test_markMain__Node() {
+    fail("Not yet implemented"); // TODO
+  } // end method */
+  
+  /**
+   * Test method for {@link ClasspathFile#markTest(Node)}.
+   */
+  @Test
+  void test_markTest__Node() {
+    // ... assertion: method checkGradleScopeTest(Node) works as expected
+    //     => not much to check hereafter
     
-    // --- c. item with "attributes" but no grand-children
-    Node child = dut.appendNode(ClasspathFile.NAME_CHILD);
-    insDut.checkGradleScopeTest(dut);
-    assertEquals(
-        "item[attributes={}; value=["
-        + "foo.bar[attributes={}; value=[]], "
-        + "attributes[attributes={}; value=[]]"
-        + "]]",
-        dut.toString()
-    );
+    // Test strategy:
+    // --- a. items with improper name are not changed
+    // --- b. items with proper name are changed
     
-    // --- d. item -> "attributes" -> "attribute" but no attributes
-    Node grand = child.appendNode(ClasspathFile.NAME_GRAND);
-    insDut.checkGradleScopeTest(dut);
-    assertEquals(
-        "item[attributes={}; value=["
-        + "foo.bar[attributes={}; value=[]], "
-        + "attributes[attributes={}; value=["
-        +   "attribute[attributes={}; value=[]]"
-        + "]]"
-        + "]]",
-        dut.toString()
-    );
-    child.remove(grand);
+    final Node root = new Node(null, "root");
+    final Map<String, String> mapProperScope = new LinkedHashMap<>();
+    mapProperScope.put("name", "gradle_used_by_scope");
+    mapProperScope.put("value", "test");
     
-    // --- e. item -> "attributes" -> "attribute" but non gradle scope attributes
-    final Map<String, String> mapE = new LinkedHashMap<>();
-    mapE.put("name", "gradle_used_by_Scope");
-    mapE.put("value", "test");
-    grand = child.appendNode(ClasspathFile.NAME_GRAND, mapE);
-    insDut.checkGradleScopeTest(dut);
-    assertEquals(
-        "item[attributes={}; value=["
-        + "foo.bar[attributes={}; value=[]], "
-        + "attributes[attributes={}; value=["
-        +   "attribute[attributes={name=gradle_used_by_Scope, value=test}; value=[]]"
-        + "]]"
-        + "]]",
-        dut.toString()
-    );
-    child.remove(grand);
+    // --- a. items with improper name are not changed
+    root.appendNode("classPathentry") // wrong capitalization
+        .appendNode(ClasspathFile.NAME_CHILD)
+        .appendNode(ClasspathFile.NAME_GRAND, mapProperScope);
     
-    // --- f. item -> "attributes" -> "attribute" but gradle scope != test
-    // f.1 gradle scope without value
-    final Map<String, String> mapF = new LinkedHashMap<>();
-    mapF.put("name", "gradle_used_by_scope");
-    grand = child.appendNode(ClasspathFile.NAME_GRAND, mapF);
-    insDut.checkGradleScopeTest(dut);
-    assertEquals(
-        "item[attributes={}; value=["
-        + "foo.bar[attributes={}; value=[]], "
-        + "attributes[attributes={}; value=["
-        +   "attribute[attributes={name=gradle_used_by_scope}; value=[]]"
-        + "]]"
-        + "]]",
-        dut.toString()
-    );
-    // f.2 gradle scope != test
-    mapF.put("value", "main,test");
-    insDut.checkGradleScopeTest(dut);
-    assertEquals(
-        "item[attributes={}; value=["
-        + "foo.bar[attributes={}; value=[]], "
-        + "attributes[attributes={}; value=["
-        +   "attribute[attributes={name=gradle_used_by_scope, value=main,test}; value=[]]"
-        + "]]"
-        + "]]",
-        dut.toString()
-    );
-    child.remove(grand);
+    // --- b. items with proper name are changed
+    root.appendNode("classpathentry")
+        .appendNode(ClasspathFile.NAME_CHILD)
+        .appendNode(ClasspathFile.NAME_GRAND, mapProperScope);
     
-    // --- g. item -> "attributes" -> "attribute", gradle scope == test, no  test-attribute
-    final String expectedG = "item[attributes={}; value=["
-        + "foo.bar[attributes={}; value=[]], "
-        + "attributes[attributes={}; value=["
-        +   "attribute[attributes={name=gradle_used_by_scope, value=test}; value=[]], "
-        +   "attribute[attributes={name=test, value=true}; value=[]]"
-        + "]]"
-        + "]]";
-    final Map<String, String> mapG = new LinkedHashMap<>();
-    mapG.put("name", "gradle_used_by_scope");
-    mapG.put("value", "test");
-    grand = child.appendNode(ClasspathFile.NAME_GRAND, mapG);
-    insDut.checkGradleScopeTest(dut);        // call method under test once
-    assertEquals(expectedG, dut.toString()); // check result
-    insDut.checkGradleScopeTest(dut);        // call method under test again
-    assertEquals(expectedG, dut.toString()); // node under test shouldn't change
-    child.remove(grand);
-    
-    // --- h. item -> "attributes" -> "attribute" but gradle scope == test plus test-attribute
-    // it is expected that method under test doesn't change node
-    dut = new Node(null, "item");
-    child = dut.appendNode(ClasspathFile.NAME_CHILD);
-    child.append(grand);
-    child.appendNode(ClasspathFile.NAME_GRAND, Map.of("name", "test"));
-    insDut.checkGradleScopeTest(dut);
+    insDut.markTest(root);
     assertEquals(
-        "item[attributes={}; value=["
-        + "attributes[attributes={}; value=["
-        +   "attribute[attributes={name=gradle_used_by_scope, value=test}; value=[]], "
-        +   "attribute[attributes={name=test}; value=[]]"
-        + "]]"
+        "root[attributes={}; value=["
+        +   "classPathentry[attributes={}; value=["
+        +     "attributes[attributes={}; value=["
+        +       "attribute[attributes={name=gradle_used_by_scope, value=test}; value=[]]"
+        +     "]]"
+        +   "]], "
+        +   "classpathentry[attributes={}; value=["
+        +     "attributes[attributes={}; value=["
+        +       "attribute[attributes={name=gradle_used_by_scope, value=test}; value=[]], "
+        +       "attribute[attributes={name=test, value=true}; value=[]]"
+        +     "]]"
+        +   "]]"
         + "]]",
-        dut.toString()
-    );
-    
-    // --- i. only first child named "attributes" count
-    // i.1 proper scope in 1st child
-    dut = new Node(null, "item");
-    final Node properScope = dut.appendNode(ClasspathFile.NAME_CHILD);
-    properScope.appendNode(ClasspathFile.NAME_GRAND, mapG);
-    insDut.checkGradleScopeTest(dut);
-    assertEquals(
-        "item[attributes={}; value=["
-        + "attributes[attributes={}; value=["
-        +   "attribute[attributes={name=gradle_used_by_scope, value=test}; value=[]], "
-        +   "attribute[attributes={name=test, value=true}; value=[]]"
-        + "]]"
-        + "]]",
-        dut.toString()
-    );
-    
-    // i.2 another improper child named "attributes" before proper child
-    dut.remove(properScope);
-    dut.appendNode(ClasspathFile.NAME_CHILD);
-    dut.appendNode(ClasspathFile.NAME_CHILD).appendNode(ClasspathFile.NAME_GRAND, mapG);
-    insDut.checkGradleScopeTest(dut);
-    assertEquals(
-        "item[attributes={}; value=["
-        + "attributes[attributes={}; value=[]], "
-        + "attributes[attributes={}; value=["
-        +   "attribute[attributes={name=gradle_used_by_scope, value=test}; value=[]]"
-        + "]]"
-        + "]]",
-        dut.toString()
+        root.toString()
     );
   } // end method */
 
+  /**
+   * Test method for {@link ClasspathFile#putJreOnModulePath(Node)}.
+   */
+  @Test
+  void test_putJreOnModulePath__Node() {
+    // Test strategy:
+    // ... assertion 1: method isJre(Node)                works as intended
+    // ... assertion 2: method hasNoAttributeModule(Node) works as intended
+    // ... assertion 3: method moveToModulePath(Node)     works as intended
+    // --- a. rootNode with a bunch of entries and entries differing slightly
+    
+    // --- a. rootNode with a bunch of entries and entries differing slightly
+    final Node rootNode = new Node(null, "root");
+    
+    // a.1: difference in name of item
+    final Map<String, String> mapA1 = new LinkedHashMap<>();
+    mapA1.put("path", "prefix" + "JRE_CONTAINER" + "suffix");
+    mapA1.put("kind", "con");
+    rootNode.appendNode("classpathentry", mapA1); // ok
+    rootNode.appendNode("classpathEntry", mapA1); // not classpathentry
+    
+    // --- improve
+    insDut.putJreOnModulePath(rootNode);
+    
+    // --- check
+    assertEquals(
+        "root[attributes={}; value=["
+        // a.1, begin
+        +   "classpathentry[attributes={path=prefixJRE_CONTAINERsuffix, kind=con}; value=["
+        +     "attributes[attributes={}; value=["
+        +       "attribute[attributes={name=module, value=true}; value=[]]"
+        +     "]]]"
+        +   "], "
+        +   "classpathEntry[attributes={path=prefixJRE_CONTAINERsuffix, kind=con}; value=[]]"
+        + "]]",
+        rootNode.toString()
+    );
+  } // end method */
+  
+  /**
+   * Test method for {@link ClasspathFile#getGradleScope(groovy.util.Node)}.
+   */
+  @Test
+  void test_getGradleScope__Node() {
+    fail("Not yet implemented"); // TODO
+  } // end method */
+  
   /**
    * Test method for {@link ClasspathFile#getAttributeNamed(Node, String)}.
    */
@@ -284,101 +263,6 @@ final class ClasspathFileTest {
     assertSame(node1, insDut.getAttributeNamed(dut, "Alfred").get());  // match in 1st proper child
     assertSame(node2, insDut.getAttributeNamed(dut, "Fiedler").get()); // match in 2nd proper child
     assertSame(node3, insDut.getAttributeNamed(dut, "bar").get());     // match in 3rd proper child
-  } // end method */
-
-  /**
-   * Test method for {@link ClasspathFile#getGradleScope(Node)}.
-   */
-  @Test
-  void test_getGradleScope__Node() {
-    // Test strategy:
-    // --- a. minimal node with gradle scope plus value
-    // --- b. minimal node with gradle scope without value
-    // --- c. minimal node without gradle scope
-    // --- d. node without appropriate children
-    // --- e. node without children
-    
-    Node dut;
-    
-    // --- a. minimal node with gradle scope plus value
-    final String valueA = "foo.bar.A";
-    final Map<String, String> mapA = new LinkedHashMap<>();
-    mapA.put("name",  "gradle_used_by_scope");
-    mapA.put("value", valueA);
-    dut = new Node(null, "item.a");
-    dut.appendNode(ClasspathFile.NAME_GRAND, mapA);
-    assertEquals(valueA, insDut.getGradleScope(dut));
-    
-    // --- b. minimal node with gradle scope without value
-    final String empty = "";
-    dut = new Node(null, "item.b");
-    dut.appendNode(ClasspathFile.NAME_GRAND, Map.of("name",  "gradle_used_by_scope"));
-    assertEquals(empty, insDut.getGradleScope(dut));
-    
-    // --- c. minimal node without gradle scope
-    dut = new Node(null, "item.c");
-    dut.appendNode(ClasspathFile.NAME_GRAND, Map.of("name",  "gradle_used_by_Scope"));
-    assertEquals(empty, insDut.getGradleScope(dut));
-    
-    // --- d. node without appropriate children
-    dut = new Node(null, "item.d");
-    dut.appendNode("foo.bar");
-    assertEquals(empty, insDut.getGradleScope(dut));
-    
-    // --- e. node without children
-    assertEquals(empty, insDut.getGradleScope(new Node(null, "item.e")));
-  } // end method */
-
-  /**
-   * Test method for {@link ClasspathFile#hasAttributeNamed(Node, String)}.
-   */
-  @Test
-  void test_hasAttributeNamed__Node_String() {
-    // Test strategy:
-    // a. Node without children SHALL return false
-    // b. Node with children not named "attribute" SHALL return false
-    // c. Node with children named "attribute" but without proper attribute SHALL return false
-    // d. Node with one or more children named "attribute" and proper attribute SHALL return true
-    
-    // --- setup a node used for testing
-    final Node dut = new Node(null, "dut"); // device under test
-    
-    // --- setup a map with attributes for child-nodes
-    final Map<String, String> mapItem = new LinkedHashMap<>();
-    mapItem.put("name", "Alfred");
-    mapItem.put("Name", "foo");
-    
-    // --- a. Node without children SHALL return false
-    assertEquals(0, dut.children().size());
-    assertFalse(insDut.hasAttributeNamed(dut, "foo"));
-
-    // --- b. Node with children not named "attribute" SHALL return false
-    dut.appendNode("Attribute", mapItem); // wrong capitalization
-    dut.appendNode("attributes", mapItem); // extra characters
-    assertFalse(insDut.hasAttributeNamed(dut, "foo"));
-    
-    // --- c. Node with children named "attribute" but without proper attribute SHALL return false
-    // c.1 child "attribute" without attributes
-    dut.appendNode(ClasspathFile.NAME_GRAND);
-    assertFalse(insDut.hasAttributeNamed(dut, "foo"));
-    
-    // c.2 child "attribute" with attributes
-    dut.appendNode(ClasspathFile.NAME_GRAND, mapItem);
-    // Note: assertions for c.2 are combined with assertions in d, see below
-    
-    // --- d. Node with children named "attribute" and proper attribute SHALL return true
-    // d.1 Just one proper child
-    assertFalse(insDut.hasAttributeNamed(dut, "alfred"));  // wrong capitalization
-    assertFalse(insDut.hasAttributeNamed(dut, " Alfred")); // extra prefix
-    assertFalse(insDut.hasAttributeNamed(dut, "Alfreds")); // extra suffix
-    assertTrue(insDut.hasAttributeNamed(dut, "Alfred"));
-    
-    // d.2 More than one proper child
-    dut.appendNode(ClasspathFile.NAME_GRAND, Map.of("name", "Fiedler"));
-    dut.appendNode(ClasspathFile.NAME_GRAND, Map.of("name", "bar"));
-    assertTrue(insDut.hasAttributeNamed(dut, "Alfred"));  // match in first  proper child
-    assertTrue(insDut.hasAttributeNamed(dut, "Fiedler")); // match in second proper child
-    assertTrue(insDut.hasAttributeNamed(dut, "bar"));     // match in third  proper child
   } // end method */
 
   /**
@@ -505,40 +389,63 @@ final class ClasspathFileTest {
   } // end method */
   
   /**
-   * Test method for {@link ClasspathFile#improveEclipseClasspathFile(Node)}.
+   * Test method for {@link ClasspathFile#hasNoAttributeTest(groovy.util.Node)}.
    */
   @Test
-  void test_improveEclipseClasspathFile__Node() {
+  void test_hasNoAttributeTest__Node() {
+    fail("Not yet implemented"); // TODO
+  } // end method */
+  
+  /**
+   * Test method for {@link ClasspathFile#hasAttributeNamed(Node, String)}.
+   */
+  @Test
+  void test_hasAttributeNamed__Node_String() {
     // Test strategy:
-    // ... assertion 1: method putJreOnModulePath(Node) works as intended
-    // --- a. JRE, rootNode with a bunch of entries and entries differing slightly
+    // a. Node without children SHALL return false
+    // b. Node with children not named "attribute" SHALL return false
+    // c. Node with children named "attribute" but without proper attribute SHALL return false
+    // d. Node with one or more children named "attribute" and proper attribute SHALL return true
     
-    // --- a. JRE, rootNode with a bunch of entries and entries differing slightly
-    final Node rootNode = new Node(null, "root");
+    // --- setup a node used for testing
+    final Node dut = new Node(null, "dut"); // device under test
     
-    // a.1: difference in name of item
-    final Map<String, String> mapA1 = new LinkedHashMap<>();
-    mapA1.put("path", "prefix" + "JRE_CONTAINER" + "suffix");
-    mapA1.put("kind", "con");
-    rootNode.appendNode("classpathentry", mapA1); // ok
-    rootNode.appendNode("classpathEntry", mapA1); // not classpathentry
+    // --- setup a map with attributes for child-nodes
+    final Map<String, String> mapItem = new LinkedHashMap<>();
+    mapItem.put("name", "Alfred");
+    mapItem.put("Name", "foo");
     
-    // --- improve
-    insDut.improveEclipseClasspathFile(rootNode);
+    // --- a. Node without children SHALL return false
+    assertEquals(0, dut.children().size());
+    assertFalse(insDut.hasAttributeNamed(dut, "foo"));
+
+    // --- b. Node with children not named "attribute" SHALL return false
+    dut.appendNode("Attribute", mapItem); // wrong capitalization
+    dut.appendNode("attributes", mapItem); // extra characters
+    assertFalse(insDut.hasAttributeNamed(dut, "foo"));
     
-    // --- check
-    assertEquals(
-        "root[attributes={}; value=["
-        // a.1, begin
-        +   "classpathentry[attributes={path=prefixJRE_CONTAINERsuffix, kind=con}; value=["
-        +     "attributes[attributes={}; value=["
-        +       "attribute[attributes={name=module, value=true}; value=[]]"
-        +     "]]]"
-        +   "], "
-        +   "classpathEntry[attributes={path=prefixJRE_CONTAINERsuffix, kind=con}; value=[]]"
-        + "]]",
-        rootNode.toString()
-    );
+    // --- c. Node with children named "attribute" but without proper attribute SHALL return false
+    // c.1 child "attribute" without attributes
+    dut.appendNode(ClasspathFile.NAME_GRAND);
+    assertFalse(insDut.hasAttributeNamed(dut, "foo"));
+    
+    // c.2 child "attribute" with attributes
+    dut.appendNode(ClasspathFile.NAME_GRAND, mapItem);
+    // Note: assertions for c.2 are combined with assertions in d, see below
+    
+    // --- d. Node with children named "attribute" and proper attribute SHALL return true
+    // d.1 Just one proper child
+    assertFalse(insDut.hasAttributeNamed(dut, "alfred"));  // wrong capitalization
+    assertFalse(insDut.hasAttributeNamed(dut, " Alfred")); // extra prefix
+    assertFalse(insDut.hasAttributeNamed(dut, "Alfreds")); // extra suffix
+    assertTrue(insDut.hasAttributeNamed(dut, "Alfred"));
+    
+    // d.2 More than one proper child
+    dut.appendNode(ClasspathFile.NAME_GRAND, Map.of("name", "Fiedler"));
+    dut.appendNode(ClasspathFile.NAME_GRAND, Map.of("name", "bar"));
+    assertTrue(insDut.hasAttributeNamed(dut, "Alfred"));  // match in first  proper child
+    assertTrue(insDut.hasAttributeNamed(dut, "Fiedler")); // match in second proper child
+    assertTrue(insDut.hasAttributeNamed(dut, "bar"));     // match in third  proper child
   } // end method */
 
   /**
@@ -689,235 +596,10 @@ final class ClasspathFileTest {
   } // end method */
   
   /**
-   * Test method for {@link ClasspathFile#markTest(Node)}.
+   * Test method for {@link ClasspathFile#move(groovy.util.Node, java.lang.String)}.
    */
   @Test
-  void test_markTest__Node() {
-    // ... assertion: method checkGradleScopeTest(Node) works as expected
-    //     => not much to check hereafter
-    
-    // Test strategy:
-    // --- a. items with improper name are not changed
-    // --- b. items with proper name are changed
-    
-    final Node root = new Node(null, "root");
-    final Map<String, String> mapProperScope = new LinkedHashMap<>();
-    mapProperScope.put("name", "gradle_used_by_scope");
-    mapProperScope.put("value", "test");
-    
-    // --- a. items with improper name are not changed
-    root.appendNode("classPathentry") // wrong capitalization
-        .appendNode(ClasspathFile.NAME_CHILD)
-        .appendNode(ClasspathFile.NAME_GRAND, mapProperScope);
-    
-    // --- b. items with proper name are changed
-    root.appendNode("classpathentry")
-        .appendNode(ClasspathFile.NAME_CHILD)
-        .appendNode(ClasspathFile.NAME_GRAND, mapProperScope);
-    
-    insDut.markTest(root);
-    assertEquals(
-        "root[attributes={}; value=["
-        +   "classPathentry[attributes={}; value=["
-        +     "attributes[attributes={}; value=["
-        +       "attribute[attributes={name=gradle_used_by_scope, value=test}; value=[]]"
-        +     "]]"
-        +   "]], "
-        +   "classpathentry[attributes={}; value=["
-        +     "attributes[attributes={}; value=["
-        +       "attribute[attributes={name=gradle_used_by_scope, value=test}; value=[]], "
-        +       "attribute[attributes={name=test, value=true}; value=[]]"
-        +     "]]"
-        +   "]]"
-        + "]]",
-        root.toString()
-    );
-  } // end method */
-
-  /**
-   * Test method for {@link ClasspathFile#moveToModulePath(Node)}.
-   */
-  @Test
-  void test_moveToModulePath__Node() {
-    // Test strategy:
-    // --- a. empty node
-    // --- b. node with children none named "attributes"
-    // --- c. node with several children named "attributes"
-    // --- d. node already moved
-    
-    Node dut;
-    
-    // --- a. empty node
-    dut = new Node(null, "rootA");
-    insDut.moveToModulePath(dut);
-    assertEquals(
-        "rootA[attributes={}; value=["
-        +   "attributes[attributes={}; value=["
-        +     "attribute[attributes={name=module, value=true}; value=[]]"
-        +   "]]"
-        + "]]",
-        dut.toString()
-    );
-    
-    // --- b. node with children none named "attributes"
-    dut = new Node(null, "rootB");
-    dut.appendNode("foo");
-    dut.appendNode("bar");
-    insDut.moveToModulePath(dut);
-    assertEquals(
-        "rootB[attributes={}; value=["
-        +   "foo[attributes={}; value=[]], "
-        +   "bar[attributes={}; value=[]], "
-        +   "attributes[attributes={}; value=["
-        +     "attribute[attributes={name=module, value=true}; value=[]]"
-        +   "]]"
-        + "]]",
-        dut.toString()
-    );
-    
-    // --- c. node with several children named "attributes"
-    // c.1 one child named "attributes"
-    dut = new Node(null, "rootC1");
-    dut.appendNode(ClasspathFile.NAME_CHILD);
-    insDut.moveToModulePath(dut);
-    assertEquals(
-        "rootC1[attributes={}; value=["
-        +   "attributes[attributes={}; value=["
-        +     "attribute[attributes={name=module, value=true}; value=[]]"
-        +   "]]"
-        + "]]",
-        dut.toString()
-    );
-    
-    // c.2 two children named "attributes" but one other node before
-    dut = new Node(null, "rootC2");
-    dut.appendNode("foo");
-    Node child = dut.appendNode(ClasspathFile.NAME_CHILD);
-    child.appendNode("bar");
-    dut.appendNode(ClasspathFile.NAME_CHILD);
-    insDut.moveToModulePath(dut);
-    assertEquals(
-        "rootC2[attributes={}; value=["
-        +   "foo[attributes={}; value=[]], "
-        +   "attributes[attributes={}; value=["
-        +     "bar[attributes={}; value=[]], "
-        +     "attribute[attributes={name=module, value=true}; value=[]]"
-        +   "]], "
-        +   "attributes[attributes={}; value=[]]"
-        + "]]",
-        dut.toString()
-    );
-    
-    // c.3 three children named "attributes" some other nodes around
-    dut = new Node(null, "rootC3");
-    dut.appendNode("foo");
-    child = dut.appendNode(ClasspathFile.NAME_CHILD, Map.of("ping", "pong"));
-    child.appendNode("bar");
-    child = dut.appendNode(ClasspathFile.NAME_CHILD);
-    child.appendNode("bar2");
-    dut.appendNode("foo2");
-    insDut.moveToModulePath(dut);
-    assertEquals(
-        "rootC3[attributes={}; value=["
-        +   "foo[attributes={}; value=[]], "
-        +   "attributes[attributes={ping=pong}; value=["
-        +     "bar[attributes={}; value=[]], "
-        +     "attribute[attributes={name=module, value=true}; value=[]]"
-        +   "]], "
-        +   "attributes[attributes={}; value=["
-        +     "bar2[attributes={}; value=[]]"
-        +   "]], "
-        +   "foo2[attributes={}; value=[]]"
-        + "]]",
-        dut.toString()
-    );
-    
-    // --- d. node already moved
-    // d.1 move information in first child with name "attributes"
-    insDut.moveToModulePath(dut);
-    assertEquals(
-        "rootC3[attributes={}; value=["
-        +   "foo[attributes={}; value=[]], "
-        +   "attributes[attributes={ping=pong}; value=["
-        +     "bar[attributes={}; value=[]], "
-        +     "attribute[attributes={name=module, value=true}; value=[]], "
-        +     "attribute[attributes={name=module, value=true}; value=[]]"
-        +   "]], "
-        +   "attributes[attributes={}; value=["
-        +     "bar2[attributes={}; value=[]]"
-        +   "]], "
-        +   "foo2[attributes={}; value=[]]"
-        + "]]",
-        dut.toString()
-    );
-    
-    // d.2 move information in 2nd child with name "attributes"
-    final Map<String, String> mapD2 = new LinkedHashMap<>();
-    mapD2.put("name", "module");
-    mapD2.put("value", "true");
-    
-    dut = new Node(null, "rootD2");
-    dut.appendNode("foo");
-    child = dut.appendNode(ClasspathFile.NAME_CHILD, Map.of("ping", "pong"));
-    child.appendNode("bar");
-    child = dut.appendNode(ClasspathFile.NAME_CHILD);
-    child.appendNode("bar2");
-    child.appendNode(ClasspathFile.NAME_GRAND, mapD2);
-    dut.appendNode("foo2");
-    insDut.moveToModulePath(dut);
-    assertEquals(
-        "rootD2[attributes={}; value=["
-        +   "foo[attributes={}; value=[]], "
-        +   "attributes[attributes={ping=pong}; value=["
-        +     "bar[attributes={}; value=[]], "
-        +     "attribute[attributes={name=module, value=true}; value=[]]"
-        +   "]], "
-        +   "attributes[attributes={}; value=["
-        +     "bar2[attributes={}; value=[]], "
-        +     "attribute[attributes={name=module, value=true}; value=[]]"
-        +   "]], "
-        +   "foo2[attributes={}; value=[]]"
-        + "]]",
-        dut.toString()
-    );
-  } // end method */
-
-  /**
-   * Test method for {@link ClasspathFile#putJreOnModulePath(Node)}.
-   */
-  @Test
-  void test_putJreOnModulePath__Node() {
-    // Test strategy:
-    // ... assertion 1: method isJre(Node)                works as intended
-    // ... assertion 2: method hasNoAttributeModule(Node) works as intended
-    // ... assertion 3: method moveToModulePath(Node)     works as intended
-    // --- a. rootNode with a bunch of entries and entries differing slightly
-    
-    // --- a. rootNode with a bunch of entries and entries differing slightly
-    final Node rootNode = new Node(null, "root");
-    
-    // a.1: difference in name of item
-    final Map<String, String> mapA1 = new LinkedHashMap<>();
-    mapA1.put("path", "prefix" + "JRE_CONTAINER" + "suffix");
-    mapA1.put("kind", "con");
-    rootNode.appendNode("classpathentry", mapA1); // ok
-    rootNode.appendNode("classpathEntry", mapA1); // not classpathentry
-    
-    // --- improve
-    insDut.putJreOnModulePath(rootNode);
-    
-    // --- check
-    assertEquals(
-        "root[attributes={}; value=["
-        // a.1, begin
-        +   "classpathentry[attributes={path=prefixJRE_CONTAINERsuffix, kind=con}; value=["
-        +     "attributes[attributes={}; value=["
-        +       "attribute[attributes={name=module, value=true}; value=[]]"
-        +     "]]]"
-        +   "], "
-        +   "classpathEntry[attributes={path=prefixJRE_CONTAINERsuffix, kind=con}; value=[]]"
-        + "]]",
-        rootNode.toString()
-    );
+  void test_move__Node_String() {
+    fail("Not yet implemented"); // TODO
   } // end method */
 } // end class
